third_party/WebKit/LayoutTests/xr/requestDevice_always_resolves.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

promise_test(t => navigator.xr.requestDevice().then(device => {
  assert_true(device == null);
}), "Test that navigator.xr.requestDevice always resolves with at least a null.");

</script>
